requirements.txt
networkx
flask
Pillow
numpy==1.14.2
opencv-python
pyyaml
easydict
scipy
h5py
invoke
boto3
xcffib
mss
progressbar2
tensorflow-gpu==1.5.0
requests
ai2thor==0.0.31
